refactor(course-page-edit): split editor zustand store
