+++ src/features/course/features/page-edit/model/editor-store/index.ts
@@ -1,0 +1,13 @@
+export { createCoursePageEditStore } from '@/features/course/features/page-edit/model/editor-store/create';
+
+export type {
+  CoursePageBlockInsertPanelState,
+  CoursePageBlockInsertPanelVisibleState,
+  CoursePageBlockInsertPlacement,
+  CoursePageEditStore,
+  CoursePageEditStoreActions,
+  CoursePageEditStoreApi,
+  CoursePageEditStoreOptions,
+  CoursePageEditStoreState,
+  CoursePageWorkingCopyUpdate,
+} from '@/features/course/features/page-edit/model/editor-store/types';
